+++ Gems/LyShine/Code/Source/UiElementComponent.cpp
@@ -142,7 +142,7 @@ void UiElementComponent::RenderElement(LyShine::IRenderGraph* renderGraph, bool 
     if (m_renderControlInterface)
     {
         // give control of rendering this element and its children to the render control component on this element
-        m_renderControlInterface->Render(renderGraph, this, m_renderInterface, m_childElementComponents.size(), isInGame);
+        m_renderControlInterface->Render(renderGraph, this, m_renderInterface, static_cast<int>(m_childElementComponents.size()), isInGame);
     }
     else
     {
@@ -153,7 +153,7 @@ void UiElementComponent::RenderElement(LyShine::IRenderGraph* renderGraph, bool 
         }
 
         // now render child elements
-        int numChildren = m_childElementComponents.size();
+        int numChildren = static_cast<int>(m_childElementComponents.size());
         for (int i = 0; i < numChildren; ++i)
         {
             GetChildElementComponent(i)->RenderElement(renderGraph, isInGame);
@@ -194,7 +194,7 @@ AZ::EntityId UiElementComponent::GetParentEntityId()
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 int UiElementComponent::GetNumChildElements()
 {
-    return m_childEntityIdOrder.size();
+    return static_cast<int>(m_childEntityIdOrder.size());
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -236,7 +236,7 @@ UiElementInterface* UiElementComponent::GetChildElementInterface(int index)
 int UiElementComponent::GetIndexOfChild(const AZ::Entity* child)
 {
     AZ::EntityId childEntityId = child->GetId();
-    int numChildren = m_childEntityIdOrder.size();
+    int numChildren = static_cast<int>(m_childEntityIdOrder.size());
     for (int i = 0;  i < numChildren; ++i)
     {
         if (m_childEntityIdOrder[i].m_entityId == childEntityId)
@@ -251,7 +251,7 @@ int UiElementComponent::GetIndexOfChild(const AZ::Entity* child)
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 int UiElementComponent::GetIndexOfChildByEntityId(AZ::EntityId childId)
 {
-    int numChildren = m_childEntityIdOrder.size();
+    int numChildren = static_cast<int>(m_childEntityIdOrder.size());
     for (int i = 0;  i < numChildren; ++i)
     {
         if (m_childEntityIdOrder[i].m_entityId == childId)
@@ -266,7 +266,7 @@ int UiElementComponent::GetIndexOfChildByEntityId(AZ::EntityId childId)
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 LyShine::EntityArray UiElementComponent::GetChildElements()
 {
-    int numChildren = m_childEntityIdOrder.size();
+    int numChildren = static_cast<int>(m_childEntityIdOrder.size());
     LyShine::EntityArray children;
     children.reserve(numChildren);
 
@@ -477,7 +477,7 @@ AZ::Entity* UiElementComponent::FindFrontmostChildContainingPoint(AZ::Vector2 po
     // this traverses all of the elements in reverse hierarchy order and returns the first one that
     // is containing the point.
     // If necessary, this could be optimized using a spatial partitioning data structure.
-    for (int i = m_childEntityIdOrder.size() - 1; !matchElem && i >= 0; i--)
+    for (int i = static_cast<int>(m_childEntityIdOrder.size() - 1); !matchElem && i >= 0; i--)
     {
         AZ::EntityId child = m_childEntityIdOrder[i].m_entityId;
 
@@ -602,7 +602,7 @@ AZ::EntityId UiElementComponent::FindInteractableToHandleEvent(AZ::Vector2 point
         EBUS_EVENT_ID_RESULT(isMasked, GetEntityId(), UiInteractionMaskBus, IsPointMasked, point);
         if (!isMasked)
         {
-            for (int i = m_childEntityIdOrder.size() - 1; !result.IsValid() && i >= 0; i--)
+            for (int i = static_cast<int>(m_childEntityIdOrder.size() - 1); !result.IsValid() && i >= 0; i--)
             {
                 result = GetChildElementComponent(i)->FindInteractableToHandleEvent(point);
             }
@@ -674,7 +674,7 @@ AZ::Entity* UiElementComponent::FindChildByName(const LyShine::NameType& name)
 
     if (AreChildPointersValid())
     {
-        int numChildren = m_childElementComponents.size();
+        int numChildren = static_cast<int>(m_childElementComponents.size());
         for (int i = 0; i < numChildren; ++i)
         {
             AZ::Entity* childEntity = GetChildElementComponent(i)->GetEntity();
@@ -709,7 +709,7 @@ AZ::Entity* UiElementComponent::FindDescendantByName(const LyShine::NameType& na
 
     if (AreChildPointersValid())
     {
-        int numChildren = m_childElementComponents.size();
+        int numChildren = static_cast<int>(m_childElementComponents.size());
         for (int i = 0; i < numChildren; ++i)
         {
             UiElementComponent* childElementComponent = GetChildElementComponent(i);
@@ -771,7 +771,7 @@ AZ::Entity* UiElementComponent::FindChildByEntityId(AZ::EntityId id)
 {
     AZ::Entity* matchElem = nullptr;
 
-    int numChildren = m_childEntityIdOrder.size();
+    int numChildren = static_cast<int>(m_childEntityIdOrder.size());
     for (int i = 0; i < numChildren; ++i)
     {
         if (id == m_childEntityIdOrder[i].m_entityId)
@@ -803,7 +803,7 @@ AZ::Entity* UiElementComponent::FindDescendantById(LyShine::ElementId id)
 
     if (AreChildPointersValid())
     {
-        int numChildren = m_childEntityIdOrder.size();
+        int numChildren = static_cast<int>(m_childEntityIdOrder.size());
         for (int i = 0; !match && i < numChildren; ++i)
         {
             match = GetChildElementComponent(i)->FindDescendantById(id);
@@ -825,7 +825,7 @@ void UiElementComponent::FindDescendantElements(AZStd::function<bool(const AZ::E
 {
     if (AreChildPointersValid())
     {
-        int numChildren = m_childElementComponents.size();
+        int numChildren = static_cast<int>(m_childElementComponents.size());
         for (int i = 0; i < numChildren; ++i)
         {
             UiElementComponent* childElementComponent = GetChildElementComponent(i);
@@ -860,7 +860,7 @@ void UiElementComponent::CallOnDescendantElements(AZStd::function<void(const AZ:
 {
     if (AreChildPointersValid())
     {
-        int numChildren = m_childEntityIdOrder.size();
+        int numChildren = static_cast<int>(m_childEntityIdOrder.size());
         for (int i = 0; i < numChildren; ++i)
         {
             callFunction(m_childEntityIdOrder[i].m_entityId);
@@ -1068,7 +1068,7 @@ void UiElementComponent::AddChild(AZ::Entity* child, AZ::Entity* insertBefore)
 
     if (insertBefore)
     {
-        int numChildren = m_childEntityIdOrder.size();
+        int numChildren = static_cast<int>(m_childEntityIdOrder.size());
         for (int i = 0; i < numChildren; ++i)
         {
             if (m_childEntityIdOrder[i].m_entityId == insertBefore->GetId())
@@ -1696,7 +1696,7 @@ void UiElementComponent::OnPatchEnd(const AZ::DataPatchNodeInfo& patchInfo)
             // the lookupAddress is the same length as the "Children" address plus an index
             // check if the address is childrenAddress plus an extra element
             bool match = true;
-            for (int i = childrenAddress.size() - 1; i >= 0; --i)
+            for (int i = static_cast<int>(childrenAddress.size() - 1); i >= 0; --i)
             {
                 if (lookupAddress[i] != childrenAddress[i])
                 {
@@ -1827,7 +1827,7 @@ void UiElementComponent::OnPatchEnd(const AZ::DataPatchNodeInfo& patchInfo)
     // This will sort all the entity order entries by sort index (primary) and entity id (secondary) which should never result in any collisions
     // This is used since slice data patching may create duplicate entries for the same sort index, missing indices and the like.
     // It should never result in multiple entity id entries since the serialization of this data uses a persistent id which is the entity id
-    int numChildren = m_childEntityIdOrder.size();
+    int numChildren = static_cast<int>(m_childEntityIdOrder.size());
     if (numChildren > 0)
     {
         AZStd::sort(m_childEntityIdOrder.begin(), m_childEntityIdOrder.end());
